chore(auth): adjust styles - clean up
